+++ packages/lumx-react/src/components/button/IconButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Emphasis, Icon, Size, Theme, Tooltip } from '@lumx/react';
+import { Emphasis, Icon, Size, Theme, Tooltip, TooltipProps } from '@lumx/react';
 import { BaseButtonProps, ButtonRoot } from '@lumx/react/components/button/ButtonRoot';
 import { COMPONENT_PREFIX } from '@lumx/react/constants';
 import { Comp, getRootClassName } from '@lumx/react/utils';
@@ -19,6 +19,8 @@ export interface IconButtonProps extends BaseButtonProps {
      * If you really don't want a tooltip and aria-label, you can give an empty label (this is not recommended).
      */
     label: string;
+    /** The props to pass to the tooltip, minus those already set by the IconButton props. */
+    tooltipProps?: Omit<TooltipProps, 'label'>;
 }
 
 /**
@@ -41,10 +43,10 @@ const DEFAULT_PROPS: Partial<IconButtonProps> = {
 };
 
 export const IconButton: Comp<IconButtonProps> = (props) => {
-    const { emphasis, icon, label, size, theme, ...forwardedProps } = props;
+    const { emphasis, icon, label, size, theme, tooltipProps, ...forwardedProps } = props;
 
     return (
-        <Tooltip label={label}>
+        <Tooltip label={label} {...tooltipProps}>
             <ButtonRoot {...{ emphasis, size, theme, ...forwardedProps }} aria-label={label} variant="icon">
                 <Icon icon={icon} />
             </ButtonRoot>