+++ packages/site-demo/content/product/components/tooltip/react/default.tsx
@@ -4,6 +4,6 @@ import React from 'react';
 
 export const App = () => (
     <Tooltip label="Print">
-        <Icon icon={mdiPrinter} />
+        <Icon icon={mdiPrinter} hasShape />
     </Tooltip>
 );